Extract ItemTextured
