+++ src/main/java/ru/mousecray/endmagic/init/EMItems.java
@@ -4,13 +4,14 @@ import net.minecraft.init.Blocks;
 import ru.mousecray.endmagic.items.EMSeeds;
 import ru.mousecray.endmagic.items.ItemNamed;
 import ru.mousecray.endmagic.items.ItemPortalBinder;
+import ru.mousecray.endmagic.items.ItemTextured;
 
 public class EMItems {
     //Not finished
     public static final EMSeeds enderSeeds = new EMSeeds(EMBlocks.enderCrops, Blocks.END_STONE, "ender_seeds", "tooltip.ender_seeds");
     public static final ItemPortalBinder itemPortalBinder = new ItemPortalBinder();
 
-    public static final ItemNamed simpletexturemodel = ItemNamed.companion.simpletexturemodelItem;
+    public static final ItemTextured simpletexturemodel = ItemTextured.companion.simpletexturemodelItem;
 
     public static final ItemNamed dragonCoal = new ItemNamed("dragon_coal");
     public static final ItemNamed naturalCoal = new ItemNamed("natural_coal");
